3 updates: a) renaming of gbooks_scrapping to gbooks_api b) update of gbooks_api c) update of covers_to_text
